+++ src/four_by_four_covermatrix.rs
@@ -1,4 +1,4 @@
-pub fn four_by_four_cover_matrix() -> Vec<Vec<usize>> {
+pub fn four_by_four_cover_matrix() -> Vec<Vec<u8>> {
     let cover_matrix = vec![
         vec![
             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
